Optimizations round 7: Length-limiting (decoder) Length-limit Huffman codes before encoding an input stream (although this is performed in the encoder, the performance improvement is in the decoder). Large Huffman codes hurts decoder performance due to the use of prefix tables; with larger codes, more bits are needed to build the table. Length-limited codes reduce this at a small cost to compression ratio.
This brings decoding time with -02 on the enwik8 file down from ~2.3 seconds to
~1.1 seconds.

+++ fly/coders/huffman_types.h
@@ -9,8 +9,7 @@ namespace fly {
 typedef std::uint8_t symbol_type;
 typedef std::uint64_t frequency_type;
 
-// TODO: Once length limting is implemented, code_type can be 16-bit.
-typedef std::uint32_t code_type;
+typedef std::uint16_t code_type;
 typedef std::uint8_t length_type;
 
 struct HuffmanNode;
